requirements.txt
playwright==1.48.0
python-telegram-bot==21.6
httpx==0.27.2
fastapi==0.115.5
uvicorn==0.32.0